requirements.txt
appdirs==1.4.4
attr==0.3.1
attrs==20.3.0
backcall==0.2.0
black==21.4b2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coremltools==4.1
cycler==0.10.0
decorator==4.4.2
elasticdeform==0.4.9
future==0.18.2
idna==2.10
imageio==2.9.0
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
kraken==2.0.8
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mpmath==1.2.1
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.8
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==300
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
sympy==1.8
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
